fix for 13988
git-svn-id: http://halodev.ontoprise.com/svn/HaloSMWExtension/trunk@10076 f80b7aa0-dfcc-4e65-a0d9-74108174cbc8

+++ extensions/SMWHalo/includes/TabularForms/TF_QP_TabularForm.php
@@ -759,7 +759,7 @@ class TFTabularFormRowData {
 			if($this->title instanceof Title){
 				$html .= $linker->makeLinkObj($this->title);
 			}
-			if($enableInstanceDelete && ($this->title instanceof Title && $this->title->exists())){
+			if($enableInstanceDelete && ($this->title instanceof Title && $this->title->exists() && $this->title->userCan('delete'))){
 				$html .= '<input class="tabf-delete-button" type="button" value="Delete" style=" display: none" onclick="tf.deleteInstance(event)"/>';
 			}
 		} else {
